requirements.txt
Django==3.0.3
djangorestframework==3.11.0
django-filter==2.3.0
djangorestframework-simplejwt==4.4.0